+++ presidio-anonymizer/tests/test_anonymizer_engine.py
@@ -16,8 +16,8 @@ from presidio_anonymizer.operators import OperatorType
 
 def test_given_request_anonymizers_return_list():
     engine = AnonymizerEngine()
-    expected_list = ["hash", "mask", "redact", "replace", "custom", "encrypt"]
-    anon_list = engine.get_anonymizers()
+    expected_list = {"hash", "mask", "redact", "replace", "custom", "keep", "encrypt"}
+    anon_list = set(engine.get_anonymizers())
 
     assert anon_list == expected_list